+++ src/services/tasks/index.ts
@@ -1,16 +1,16 @@
 import type { HttpInstance } from '../../types';
 import httpClient from '../http-client';
-import tasksService from './tasksService';
+import tasks from './tasks';
 import { TASKS_BASE } from '../../constants';
 
-export type TasksService = ReturnType<typeof tasksService>;
+export type TasksService = ReturnType<typeof tasks>;
 
-export default (httpWithAuth: HttpInstance): TasksService => {
+export const tasksService = (httpWithAuth: HttpInstance): TasksService => {
   const client = httpClient({
     basePath: TASKS_BASE,
   });
 
-  const tasksMethods = tasksService(client, httpWithAuth);
+  const tasksMethods = tasks(client, httpWithAuth);
 
   return {
     ...tasksMethods,